+++ Cargo.toml
@@ -26,17 +26,23 @@ alloc = []
 log = ["dep:log"]
 
 # Enable support for the `serial2` crate.
-serial2 = ["dep:serial2", "std"]
+serial2 = ["dep:serial2", "std", "serial2-tokio"]
 
-# Enable rs4xx support of the `serial2` crate.
-rs4xx = ["serial2?/rs4xx"]
+# Enable support for the `serial2-tokio` crate.
+serial2-tokio = ["dep:serial2-tokio", "tokio", "std"]
+
+# Enable rs4xx support of the `serial2` and `serial2-tokio` crate.
+rs4xx = ["serial2?/rs4xx", "serial2-tokio?/rs4xx"]
 
 # Enable integration tests that use a real servo motor.
 integration-tests = []
 
 [dependencies]
+bisync = "0.3.0"
 log = { version = "0.4.8", optional = true }
 serial2 = { version = "0.2.24", optional = true }
+serial2-tokio = { version = "0.1.19", optional = true }
+tokio = { version = "1.48.0", features = ["time"], optional = true }
 
 [dev-dependencies]
 assert2 = "0.3.3"
