+++ src/bus/mod.rs
@@ -1,28 +1,7 @@
 //! Low level interface to a DYNAMIXEL Protocol 2.0 bus.
 
+use crate::bus_types::{bytestuff, endian, InstructionPacket, Packet, StatusPacket, HEADER_PREFIX, HEADER_SIZE};
 use crate::{checksum, ReadError, WriteError};
-use core::time::Duration;
-
-pub(crate) mod bytestuff;
-pub(crate) mod endian;
-
-pub(crate) mod data;
-pub use data::Data;
-
-mod packet;
-pub use packet::{InstructionPacket, Packet, StatusPacket};
-
-/// Prefix of a packet.
-///
-/// All packets start with this prefix, and they can not contain it in the body.
-///
-/// In other words: if you see this in the data stream, it must be the start of a packet.
-const HEADER_PREFIX: [u8; 4] = [0xFF, 0xFF, 0xFD, 0x00];
-
-/// The size of a message header, including the pre-amble, packet ID and length.
-///
-/// Excludes the instruction ID, the error field of status packets, the parameters and the CRC.
-const HEADER_SIZE: usize = 7;
 
 /// Default buffer type.
 ///
@@ -38,49 +17,13 @@ pub type DefaultBuffer = alloc::vec::Vec<u8>;
 #[cfg(not(feature = "alloc"))]
 pub type DefaultBuffer = &'static mut [u8];
 
-/// Create a mutable static buffer of size N.
-///
-/// This macro returns a `&'mut static [u8]`.
-/// Each occurence of the macro may only be evaluated once,
-/// any subsequent invocation will panic.
-///
-/// The macro is usable as expression in const context.
-///
-/// # Usage:
-/// ```no_run
-/// # fn main() -> Result<(), std::io::Error> {
-/// # let serial_port = serial2::SerialPort::open("/dev/null", 57600)?;
-/// use dynamixel2::{Client, static_buffer};
-/// let client = Client::with_buffers(serial_port, static_buffer!(128), static_buffer!(64))?;
-/// # Ok(())
-/// # }
-/// ```
-#[macro_export]
-macro_rules! static_buffer {
-	($N:literal) => {{
-		use ::core::sync::atomic::{AtomicBool, Ordering};
-		static USED: AtomicBool = AtomicBool::new(false);
-		static mut BUFFER: [u8; $N] = [0; $N];
-		if USED.swap(true, Ordering::Relaxed) {
-			panic!("static buffer already used, each occurence of `static_buffer!()` may only be evaluated once");
-		}
-		unsafe {
-			// Use raw pointer to avoid compiler warning.
-			let buffer = &raw mut BUFFER;
-			// Convert to reference in separate expression to avoid clippy warning.
-			let buffer = &mut *buffer;
-			buffer.as_mut_slice()
-		}
-	}};
-}
-
 /// Low level interface to a DYNAMIXEL Protocol 2.0 bus.
 ///
 /// Does not assume anything about the direction of communication.
 /// Used by [`crate::Client`] and [`crate::Device`].
 pub(crate) struct Bus<SerialPort, Buffer>
 where
-	SerialPort: crate::SerialPort,
+	SerialPort: super::SerialPort,
 	Buffer: AsRef<[u8]> + AsMut<[u8]>,
 {
 	/// The underlying stream (normally a serial port).
@@ -102,9 +45,10 @@ where
 	pub(crate) write_buffer: Buffer,
 }
 
+#[super::bisync]
 impl<SerialPort, Buffer> Bus<SerialPort, Buffer>
 where
-	SerialPort: crate::SerialPort,
+	SerialPort: super::SerialPort,
 	Buffer: AsRef<[u8]> + AsMut<[u8]>,
 {
 	/// Create a new bus using pre-allocated buffers.
@@ -143,7 +87,7 @@ where
 	}
 
 	/// Write a status message to the bus.
-	pub fn write_status<F>(
+	pub async fn write_status<F>(
 		&mut self,
 		packet_id: u8,
 		error: u8,
@@ -154,19 +98,15 @@ where
 		F: FnOnce(&mut [u8]) -> Result<(), crate::error::BufferTooSmallError>,
 	{
 		crate::error::BufferTooSmallError::check(StatusPacket::message_len(parameter_count), self.write_buffer.as_ref().len())?;
-		self.write_packet(
-			packet_id,
-			crate::instructions::instruction_id::STATUS,
-			parameter_count + 1,
-			|buffer| {
-				buffer[0] = error;
-				encode_parameters(&mut buffer[1..])
-			},
-		)
+		self.write_packet(packet_id, crate::instruction_id::STATUS, parameter_count + 1, |buffer| {
+			buffer[0] = error;
+			encode_parameters(&mut buffer[1..])
+		})
+		.await
 	}
 
 	/// Write an instruction message to the bus.
-	pub fn write_instruction<F>(
+	pub async fn write_instruction<F>(
 		&mut self,
 		packet_id: u8,
 		instruction_id: u8,
@@ -177,10 +117,11 @@ where
 		F: FnOnce(&mut [u8]) -> Result<(), crate::error::BufferTooSmallError>,
 	{
 		self.write_packet(packet_id, instruction_id, parameter_count, encode_parameters)
+			.await
 	}
 
 	/// Write a packet to the bus.
-	pub fn write_packet<F>(
+	pub async fn write_packet<F>(
 		&mut self,
 		packet_id: u8,
 		instruction_id: u8,
@@ -224,12 +165,12 @@ where
 		// Send message.
 		let stuffed_message = &buffer[..checksum_index + 2];
 		trace!("sending packet: {:02X?}", stuffed_message);
-		self.serial_port.write_all(stuffed_message).map_err(WriteError::Write)?;
+		self.serial_port.write_all(stuffed_message).await.map_err(WriteError::Write)?;
 		Ok(())
 	}
 
 	/// Read a raw packet from the bus with the given deadline.
-	pub fn read_packet_deadline(&mut self, deadline: SerialPort::Instant) -> Result<Packet<'_>, ReadError<SerialPort::Error>> {
+	pub async fn read_packet_deadline(&mut self, deadline: SerialPort::Instant) -> Result<Packet<'_>, ReadError<SerialPort::Error>> {
 		// Check that the read buffer is large enough to hold atleast a instruction packet with 0 parameters.
 		crate::error::BufferTooSmallError::check(HEADER_SIZE + 3, self.read_buffer.as_mut().len())?;
 
@@ -256,6 +197,7 @@ where
 			let new_data = self
 				.serial_port
 				.read(&mut self.read_buffer.as_mut()[self.read_len..], &deadline)
+				.await
 				.map_err(ReadError::Io)?;
 
 			self.read_len += new_data;
@@ -284,10 +226,10 @@ where
 
 		// Wrap the data in a `Packet`.
 		let data = &self.read_buffer.as_ref()[..HEADER_SIZE + parameter_count];
-		let packet = packet::Packet { data };
+		let packet = Packet { data };
 
 		// Ensure that status packets have an error field (included in parameter_count here).
-		if packet.instruction_id() == crate::instructions::instruction_id::STATUS && parameter_count < 1 {
+		if packet.instruction_id() == crate::instruction_id::STATUS && parameter_count < 1 {
 			return Err(crate::InvalidMessage::InvalidParameterCount(crate::InvalidParameterCount {
 				actual: 0,
 				expected: crate::ExpectedCount::Min(1),
@@ -337,70 +279,13 @@ fn find_header(buffer: &[u8]) -> usize {
 	buffer.len()
 }
 
-/// Calculate the required time to transfer a message of a given size.
-///
-/// The size must include any headers and footers of the message.
-pub(crate) fn message_transfer_time(message_size: u32, baud_rate: u32) -> Duration {
-	let baud_rate = u64::from(baud_rate);
-	let bits = u64::from(message_size) * 10; // each byte is 1 start bit, 8 data bits and 1 stop bit.
-	let secs = bits / baud_rate;
-	let subsec_bits = bits % baud_rate;
-	let nanos = (subsec_bits * 1_000_000_000).div_ceil(baud_rate);
-	Duration::new(secs, nanos as u32)
-}
-
 #[cfg(test)]
 mod test {
+	use core::time::Duration;
+
 	use super::*;
+	use crate::static_buffer;
 	use assert2::assert;
-
-	#[test]
-	fn test_message_transfer_time() {
-		// Try a bunch of values to ensure we dealt with overflow correctly.
-		assert!(message_transfer_time(100, 1_000) == Duration::from_secs(1));
-		assert!(message_transfer_time(1_000, 10_000) == Duration::from_secs(1));
-		assert!(message_transfer_time(1_000, 1_000_000) == Duration::from_millis(10));
-		assert!(message_transfer_time(1_000, 10_000_000) == Duration::from_millis(1));
-		assert!(message_transfer_time(1_000, 100_000_000) == Duration::from_micros(100));
-		assert!(message_transfer_time(1_000, 1_000_000_000) == Duration::from_micros(10));
-		assert!(message_transfer_time(1_000, 2_000_000_000) == Duration::from_micros(5));
-		assert!(message_transfer_time(1_000, 4_000_000_000) == Duration::from_nanos(2500));
-		assert!(message_transfer_time(10_000, 4_000_000_000) == Duration::from_micros(25));
-		assert!(message_transfer_time(1_000_000, 4_000_000_000) == Duration::from_micros(2500));
-		assert!(message_transfer_time(10_000_000, 4_000_000_000) == Duration::from_millis(25));
-		assert!(message_transfer_time(100_000_000, 4_000_000_000) == Duration::from_millis(250));
-		assert!(message_transfer_time(1_000_000_000, 4_000_000_000) == Duration::from_millis(2500));
-		assert!(message_transfer_time(2_000_000_000, 4_000_000_000) == Duration::from_secs(5));
-		assert!(message_transfer_time(4_000_000_000, 4_000_000_000) == Duration::from_secs(10));
-		assert!(message_transfer_time(4_000_000_000, 2_000_000_000) == Duration::from_secs(20));
-		assert!(message_transfer_time(4_000_000_000, 1_000_000_000) == Duration::from_secs(40));
-		assert!(message_transfer_time(4_000_000_000, 100_000_000) == Duration::from_secs(400));
-		assert!(message_transfer_time(4_000_000_000, 10_000_000) == Duration::from_secs(4_000));
-		assert!(message_transfer_time(4_000_000_000, 1_000_000) == Duration::from_secs(40_000));
-		assert!(message_transfer_time(4_000_000_000, 100_000) == Duration::from_secs(400_000));
-		assert!(message_transfer_time(4_000_000_000, 10_000) == Duration::from_secs(4_000_000));
-		assert!(message_transfer_time(4_000_000_000, 1_000) == Duration::from_secs(40_000_000));
-		assert!(message_transfer_time(4_000_000_000, 100) == Duration::from_secs(400_000_000));
-		assert!(message_transfer_time(4_000_000_000, 10) == Duration::from_secs(4_000_000_000));
-		assert!(message_transfer_time(4_000_000_000, 1) == Duration::from_secs(40_000_000_000));
-
-		assert!(message_transfer_time(43, 1) == Duration::from_secs(430));
-		assert!(message_transfer_time(43, 10) == Duration::from_secs(43));
-		assert!(message_transfer_time(43, 2) == Duration::from_secs(215));
-		assert!(message_transfer_time(43, 20) == Duration::from_millis(21_500));
-		assert!(message_transfer_time(43, 200) == Duration::from_millis(2_150));
-		assert!(message_transfer_time(43, 2_000_000) == Duration::from_micros(215));
-		assert!(message_transfer_time(43, 2_000_000_000) == Duration::from_nanos(215));
-		assert!(message_transfer_time(43, 4_000_000_000) == Duration::from_nanos(108)); // rounded up
-		assert!(message_transfer_time(3, 4_000_000_000) == Duration::from_nanos(8)); // rounded up
-		assert!(message_transfer_time(5, 4_000_000_000) == Duration::from_nanos(13)); // rounded up
-
-		let lots = u32::MAX - 1; // Use MAX - 1 because MAX is not cleanly divisible by 2.
-		assert!(message_transfer_time(lots, 1) == Duration::from_secs(u64::from(lots) * 10));
-		assert!(message_transfer_time(lots, lots) == Duration::from_secs(10));
-		assert!(message_transfer_time(lots / 2, lots) == Duration::from_secs(5));
-		assert!(message_transfer_time(lots, lots / 2) == Duration::from_secs(20));
-	}
 
 	#[test]
 	fn test_find_garbage_end() {
@@ -420,48 +305,11 @@ mod test {
 		assert!(find_header(&[0, 1, 2, 3, 4, 0xFF, 6]) == 7);
 	}
 
-	#[test]
-	#[allow(clippy::needless_range_loop, reason = "indexing gives better error messsage on failed assertion")]
-	fn test_static_buffer() {
-		let buffer1 = static_buffer!(128);
-		assert!(buffer1.len() == 128);
-		for i in 0..buffer1.len() {
-			assert!(buffer1[i] == 0, "i: {i}");
-		}
-
-		let buffer2 = static_buffer!(64);
-		assert!(buffer2.len() == 64);
-		for i in 0..buffer2.len() {
-			assert!(buffer2[i] == 0, "i: {i}");
-		}
-	}
-
-	#[test]
-	#[should_panic]
-	fn test_static_buffer_panics_when_evaluated_in_loop() {
-		for _ in 0..2 {
-			let buffer = static_buffer!(128);
-			assert!(buffer.len() == 128);
-		}
-	}
-
-	#[test]
-	#[should_panic]
-	fn test_static_buffer_panics_when_evaluated_in_loop_through_function() {
-		fn make_buffer() -> &'static mut [u8] {
-			static_buffer!(128)
-		}
-
-		for _ in 0..2 {
-			let buffer = make_buffer();
-			assert!(buffer.len() == 128);
-		}
-	}
-
+	#[super::super::only_sync]
 	#[test]
 	fn test_buffer_too_small() {
-		let read_buffer = static_buffer!(128);
-		let write_buffer = static_buffer!(128);
+		let read_buffer = crate::static_buffer!(128);
+		let write_buffer = crate::static_buffer!(128);
 
 		// Dummy serial port used to feed packages
 		// It does not support writing them etc.
@@ -508,7 +356,7 @@ mod test {
 				unimplemented!("not used in this test")
 			}
 
-			fn make_deadline(&self, _timeout: Duration) -> Self::Instant {
+			fn make_deadline(&self, _timeout: core::time::Duration) -> Self::Instant {
 				std::time::Instant::now() + _timeout
 			}
