+++ src/bus_types/data.rs
@@ -1,4 +1,5 @@
-use crate::bus::StatusPacket;
+use crate::bus_types::StatusPacket;
+
 use crate::Response;
 use core::mem::MaybeUninit;
