+++ src/client.rs
@@ -1,8 +1,7 @@
+use super::bus::Bus;
+use crate::instruction_id;
+use crate::{bus_types::StatusPacket, ReadError, TransferError, WriteError};
 use core::time::Duration;
-
-use crate::bus::{Bus, StatusPacket};
-use crate::instructions::instruction_id;
-use crate::{ReadError, TransferError, WriteError};
 
 macro_rules! make_client_struct {
 	($($DefaultSerialPort:ty)?) => {
@@ -18,7 +17,7 @@ macro_rules! make_client_struct {
 		/// See the [`crate::static_buffer!()`] macro for a way to safely create a mutable static buffer.
 		pub struct Client<SerialPort $(= $DefaultSerialPort)?, Buffer = crate::bus::DefaultBuffer>
 		where
-			SerialPort: crate::SerialPort,
+			SerialPort: super::SerialPort,
 			Buffer: AsRef<[u8]> + AsMut<[u8]>,
 		{
 			bus: Bus<SerialPort, Buffer>,
@@ -27,14 +26,14 @@ macro_rules! make_client_struct {
 }
 
 #[cfg(feature = "serial2")]
-make_client_struct!(serial2::SerialPort);
+make_client_struct!(super::Serial2Port);
 
 #[cfg(not(feature = "serial2"))]
 make_client_struct!();
 
 impl<SerialPort, Buffer> core::fmt::Debug for Client<SerialPort, Buffer>
 where
-	SerialPort: crate::SerialPort + core::fmt::Debug,
+	SerialPort: super::SerialPort + core::fmt::Debug,
 	Buffer: AsRef<[u8]> + AsMut<[u8]>,
 {
 	fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
@@ -46,20 +45,20 @@ where
 }
 
 #[cfg(feature = "serial2")]
-impl Client<serial2::SerialPort, Vec<u8>> {
+impl Client<super::Serial2Port, Vec<u8>> {
 	/// Open a serial port with the given baud rate.
 	///
 	/// This will allocate a new read and write buffer of 128 bytes each.
 	/// Use [`Self::open_with_buffers()`] if you want to use a custom buffers.
 	pub fn open(path: impl AsRef<std::path::Path>, baud_rate: u32) -> std::io::Result<Self> {
-		let serial_port = serial2::SerialPort::open(path, baud_rate)?;
+		let serial_port = super::Serial2Port::open(path, baud_rate)?;
 		let bus = Bus::with_buffers_and_baud_rate(serial_port, vec![0; 128], vec![0; 128], baud_rate);
 		Ok(Self { bus })
 	}
 }
 
 #[cfg(feature = "serial2")]
-impl<Buffer> Client<serial2::SerialPort, Buffer>
+impl<Buffer> Client<super::Serial2Port, Buffer>
 where
 	Buffer: AsRef<[u8]> + AsMut<[u8]>,
 {
@@ -72,7 +71,7 @@ where
 		read_buffer: Buffer,
 		write_buffer: Buffer,
 	) -> std::io::Result<Self> {
-		let serial_port = serial2::SerialPort::open(path, baud_rate)?;
+		let serial_port = super::Serial2Port::open(path, baud_rate)?;
 		let bus = Bus::with_buffers_and_baud_rate(serial_port, read_buffer, write_buffer, baud_rate);
 		Ok(Self { bus })
 	}
@@ -81,7 +80,7 @@ where
 #[cfg(feature = "alloc")]
 impl<SerialPort> Client<SerialPort, alloc::vec::Vec<u8>>
 where
-	SerialPort: crate::SerialPort,
+	SerialPort: super::SerialPort,
 {
 	/// Create a new client using an open serial port.
 	///
@@ -97,9 +96,10 @@ where
 	}
 }
 
+#[super::bisync]
 impl<SerialPort, Buffer> Client<SerialPort, Buffer>
 where
-	SerialPort: crate::SerialPort,
+	SerialPort: super::SerialPort,
 	Buffer: AsRef<[u8]> + AsMut<[u8]>,
 {
 	/// Create a new client using pre-allocated buffers.
@@ -146,7 +146,7 @@ where
 	/// This is not suitable for broadcast instructions.
 	/// For broadcast instructions, each motor sends an individual response or no response is send at all.
 	/// Instead, use [`Self::write_instruction`] and [`Self::read_status_response`].
-	pub fn transfer_single<F>(
+	pub async fn transfer_single<F>(
 		&mut self,
 		packet_id: u8,
 		instruction_id: u8,
@@ -157,14 +157,15 @@ where
 	where
 		F: FnOnce(&mut [u8]) -> Result<(), crate::error::BufferTooSmallError>,
 	{
-		self.write_instruction(packet_id, instruction_id, parameter_count, encode_parameters)?;
-		let response = self.read_status_response(expected_response_parameters)?;
+		self.write_instruction(packet_id, instruction_id, parameter_count, encode_parameters)
+			.await?;
+		let response = self.read_status_response(expected_response_parameters).await?;
 		crate::error::InvalidPacketId::check(response.packet_id(), packet_id).map_err(crate::ReadError::from)?;
 		Ok(response)
 	}
 
 	/// Write an instruction message to the bus.
-	pub fn write_instruction<F>(
+	pub async fn write_instruction<F>(
 		&mut self,
 		packet_id: u8,
 		instruction_id: u8,
@@ -176,12 +177,16 @@ where
 	{
 		self.bus
 			.write_instruction(packet_id, instruction_id, parameter_count, encode_parameters)
+			.await
 	}
 
 	/// Read a raw status response from the bus with the given deadline.
-	pub fn read_status_response_timeout<'a>(&'a mut self, timeout: Duration) -> Result<StatusPacket<'a>, ReadError<SerialPort::Error>> {
+	pub async fn read_status_response_timeout<'a>(
+		&'a mut self,
+		timeout: Duration,
+	) -> Result<StatusPacket<'a>, ReadError<SerialPort::Error>> {
 		let deadline = self.serial_port().make_deadline(timeout);
-		let packet = self.bus.read_packet_deadline(deadline)?;
+		let packet = self.bus.read_packet_deadline(deadline).await?;
 		let status = match packet.as_status() {
 			Some(status) => status,
 			None => {
@@ -200,10 +205,13 @@ where
 	/// Read a raw status response with an automatically calculated timeout.
 	///
 	/// The read timeout is determined by the expected number of response parameters and the baud rate of the bus.
-	pub fn read_status_response<'a>(&'a mut self, expected_parameters: u16) -> Result<StatusPacket<'a>, ReadError<SerialPort::Error>> {
+	pub async fn read_status_response<'a>(
+		&'a mut self,
+		expected_parameters: u16,
+	) -> Result<StatusPacket<'a>, ReadError<SerialPort::Error>> {
 		// Official SDK adds a flat 34 milliseconds, so lets just mimick that.
-		let message_size = crate::bus::StatusPacket::message_len(expected_parameters as usize) as u32;
-		let timeout = crate::bus::message_transfer_time(message_size, self.bus.baud_rate) + Duration::from_millis(34);
-		self.read_status_response_timeout(timeout)
+		let message_size = crate::bus_types::StatusPacket::message_len(expected_parameters as usize) as u32;
+		let timeout = crate::bus_types::message_transfer_time(message_size, self.bus.baud_rate) + Duration::from_millis(34);
+		self.read_status_response_timeout(timeout).await
 	}
 }
