+++ src/instructions/action.rs
@@ -1,9 +1,11 @@
-use super::{instruction_id, packet_id};
-use crate::{Client, Response, TransferError, WriteError};
+use super::Client;
+use crate::{instruction_id, packet_id};
+use crate::{Response, TransferError, WriteError};
 
+#[super::bisync]
 impl<SerialPort, Buffer> Client<SerialPort, Buffer>
 where
-	SerialPort: crate::SerialPort,
+	SerialPort: super::SerialPort,
 	Buffer: AsRef<[u8]> + AsMut<[u8]>,
 {
 	/// Send an action command to trigger a previously registered instruction.
@@ -12,13 +14,14 @@ where
 	/// If you do, none of the devices will reply with a response, and this function will not wait for any.
 	///
 	/// If you want to broadcast this instruction, it may be more convenient to use [`Self::broadcast_action()`] instead.
-	pub fn action(&mut self, motor_id: u8) -> Result<Response<()>, TransferError<SerialPort::Error>> {
-		self.write_instruction(motor_id, instruction_id::ACTION, 0, |_| Ok(()))?;
-		Ok(super::read_response_if_not_broadcast(self, motor_id)?)
+	pub async fn action(&mut self, motor_id: u8) -> Result<Response<()>, TransferError<SerialPort::Error>> {
+		self.write_instruction(motor_id, instruction_id::ACTION, 0, |_| Ok(())).await?;
+		Ok(super::read_response_if_not_broadcast(self, motor_id).await?)
 	}
 
 	/// Broadcast an action command to all connected motors to trigger a previously registered instruction.
-	pub fn broadcast_action(&mut self) -> Result<(), WriteError<SerialPort::Error>> {
+	pub async fn broadcast_action(&mut self) -> Result<(), WriteError<SerialPort::Error>> {
 		self.write_instruction(packet_id::BROADCAST, instruction_id::ACTION, 0, |_| Ok(()))
+			.await
 	}
 }
