+++ src/instructions/reg_write.rs
@@ -1,11 +1,14 @@
-use super::{instruction_id, read_response_if_not_broadcast};
-use crate::bus::endian::write_u16_le;
-use crate::bus::Data;
-use crate::{Client, Response, TransferError};
+use super::read_response_if_not_broadcast;
+use super::Client;
+use crate::bus_types::endian::write_u16_le;
+use crate::bus_types::Data;
+use crate::instruction_id;
+use crate::{Response, TransferError};
 
+#[super::bisync]
 impl<SerialPort, Buffer> Client<SerialPort, Buffer>
 where
-	SerialPort: crate::SerialPort,
+	SerialPort: super::SerialPort,
 	Buffer: AsRef<[u8]> + AsMut<[u8]>,
 {
 	/// Register a write of an arbitrary number of bytes, to be triggered later by an `action` command.
@@ -17,13 +20,19 @@ where
 	///
 	/// You may specify [`crate::instructions::packet_id::BROADCAST`] as motor ID.
 	/// If you do, none of the devices will reply with a response, and this function will not wait for any.
-	pub fn reg_write_bytes(&mut self, motor_id: u8, address: u16, data: &[u8]) -> Result<Response<()>, TransferError<SerialPort::Error>> {
+	pub async fn reg_write_bytes(
+		&mut self,
+		motor_id: u8,
+		address: u16,
+		data: &[u8],
+	) -> Result<Response<()>, TransferError<SerialPort::Error>> {
 		self.write_instruction(motor_id, instruction_id::REG_WRITE, 2 + data.len(), |buffer| {
 			write_u16_le(&mut buffer[0..], address);
 			buffer[2..].copy_from_slice(data);
 			Ok(())
-		})?;
-		Ok(read_response_if_not_broadcast(self, motor_id)?)
+		})
+		.await?;
+		Ok(read_response_if_not_broadcast(self, motor_id).await?)
 	}
 
 	/// Register a write command for value to a specific motor.
@@ -35,11 +44,17 @@ where
 	///
 	/// You may specify [`crate::instructions::packet_id::BROADCAST`] as motor ID.
 	/// If you do, none of the devices will reply with a response, and this function will not wait for any.
-	pub fn reg_write<T: Data>(&mut self, motor_id: u8, address: u16, value: &T) -> Result<Response<()>, TransferError<SerialPort::Error>> {
+	pub async fn reg_write<T: Data>(
+		&mut self,
+		motor_id: u8,
+		address: u16,
+		value: &T,
+	) -> Result<Response<()>, TransferError<SerialPort::Error>> {
 		self.write_instruction(motor_id, instruction_id::REG_WRITE, 2 + 1, |buffer| {
 			write_u16_le(&mut buffer[0..], address);
 			value.encode(&mut buffer[2..])
-		})?;
-		Ok(read_response_if_not_broadcast(self, motor_id)?)
+		})
+		.await?;
+		Ok(read_response_if_not_broadcast(self, motor_id).await?)
 	}
 }
