+++ src/instructions/sync_write.rs
@@ -1,10 +1,12 @@
-use super::{instruction_id, packet_id, SyncWriteData};
-use crate::bus::endian::write_u16_le;
-use crate::{Client, WriteError};
+use super::Client;
+use crate::bus_types::endian::write_u16_le;
+use crate::WriteError;
+use crate::{instruction_id, packet_id, SyncWriteData};
 
+#[super::bisync]
 impl<SerialPort, Buffer> Client<SerialPort, Buffer>
 where
-	SerialPort: crate::SerialPort,
+	SerialPort: super::SerialPort,
 	Buffer: AsRef<[u8]> + AsMut<[u8]>,
 {
 	/// Synchronously write an arbitrary number of bytes to multiple motors.
@@ -38,7 +40,7 @@ where
 	/// # Ok(())
 	/// # }
 	/// ```
-	pub fn sync_write_bytes<'a, Iter, Data, Buf>(
+	pub async fn sync_write_bytes<'a, Iter, Data, Buf>(
 		&mut self,
 		address: u16,
 		count: u16,
@@ -66,18 +68,19 @@ where
 			}
 			Ok(())
 		})
+		.await
 	}
 
 	/// Synchronously write a value to multiple motors.
 	///
 	/// Each motor will perform the write as soon as it receives the command.
 	/// This gives much shorter delays than executing a regular [`Self::write`] for each motor individually.
-	pub fn sync_write<Iter, Data, T>(&mut self, address: u16, data: Iter) -> Result<(), WriteError<SerialPort::Error>>
+	pub async fn sync_write<Iter, Data, T>(&mut self, address: u16, data: Iter) -> Result<(), WriteError<SerialPort::Error>>
 	where
 		Iter: IntoIterator<Item = Data>,
 		Iter::IntoIter: ExactSizeIterator,
 		Data: AsRef<SyncWriteData<T>>,
-		T: crate::bus::Data,
+		T: crate::bus_types::Data,
 	{
 		let data = data.into_iter();
 		let count = T::ENCODED_SIZE;
@@ -95,5 +98,6 @@ where
 			}
 			Ok(())
 		})
+		.await
 	}
 }
