+++ src/instructions/write.rs
@@ -1,36 +1,41 @@
-use super::{instruction_id, read_response_if_not_broadcast};
-use crate::bus::endian::write_u16_le;
-use crate::bus::Data;
-use crate::{Client, Response, TransferError};
+use super::read_response_if_not_broadcast;
+use super::Client;
+use crate::bus_types::endian::write_u16_le;
+use crate::bus_types::Data;
+use crate::instruction_id;
+use crate::{Response, TransferError};
 
+#[super::bisync]
 impl<SerialPort, Buffer> Client<SerialPort, Buffer>
 where
-	SerialPort: crate::SerialPort,
+	SerialPort: super::SerialPort,
 	Buffer: AsRef<[u8]> + AsMut<[u8]>,
 {
 	/// Write value to a specific motor.
 	///
 	/// You may specify [`crate::instructions::packet_id::BROADCAST`] as motor ID.
 	/// If you do, none of the devices will reply with a response, and this function will not wait for any.
-	pub fn write<T: Data>(&mut self, motor_id: u8, address: u16, data: &T) -> Result<Response<()>, TransferError<SerialPort::Error>> {
+	pub async fn write<T: Data>(&mut self, motor_id: u8, address: u16, data: &T) -> Result<Response<()>, TransferError<SerialPort::Error>> {
 		self.write_instruction(motor_id, instruction_id::WRITE, 2 + T::ENCODED_SIZE as usize, |buffer| {
 			write_u16_le(&mut buffer[0..], address);
 			data.encode(&mut buffer[2..])?;
 			Ok(())
-		})?;
-		Ok(read_response_if_not_broadcast(self, motor_id)?)
+		})
+		.await?;
+		Ok(read_response_if_not_broadcast(self, motor_id).await?)
 	}
 
 	/// Write an arbitrary amount of bytes to a specific motor.
 	///
 	/// You may specify [`crate::instructions::packet_id::BROADCAST`] as motor ID.
 	/// If you do, none of the devices will reply with a response, and this function will not wait for any.
-	pub fn write_bytes(&mut self, motor_id: u8, address: u16, data: &[u8]) -> Result<Response<()>, TransferError<SerialPort::Error>> {
+	pub async fn write_bytes(&mut self, motor_id: u8, address: u16, data: &[u8]) -> Result<Response<()>, TransferError<SerialPort::Error>> {
 		self.write_instruction(motor_id, instruction_id::WRITE, 2 + data.len(), |buffer| {
 			write_u16_le(&mut buffer[0..], address);
 			buffer[2..].copy_from_slice(data);
 			Ok(())
-		})?;
-		Ok(read_response_if_not_broadcast(self, motor_id)?)
+		})
+		.await?;
+		Ok(read_response_if_not_broadcast(self, motor_id).await?)
 	}
 }
