+++ src/lib.rs
@@ -26,14 +26,54 @@
 #[cfg(feature = "alloc")]
 extern crate alloc;
 
-#[cfg(feature = "serial2")]
-/// Public re-export of the serial2 crate.
-pub use serial2;
-
 #[macro_use]
 mod log;
 
 pub mod checksum;
+
+pub(crate) mod bus_types;
+pub use bus_types::*;
+
+mod common;
+pub use common::*;
+
+#[path = "."]
+/// asynchronous dynamixel2 Bus
+pub mod asynch {
+	use bisync::asynchronous::*;
+
+	#[cfg(feature = "serial2-tokio")]
+	/// Public re-export of the serial2 crate.
+	pub use serial2_tokio;
+	#[cfg(feature = "serial2-tokio")]
+	use serial2_tokio::SerialPort as Serial2Port;
+	pub mod instructions;
+
+	mod client;
+	pub use client::*;
+
+	mod device;
+	pub use device::*;
+
+	mod serial_port;
+	pub use serial_port::SerialPort;
+
+	// mod error;
+	// pub use error::*;
+
+	// mod response;
+	// pub use response::*;
+
+	pub mod bus;
+}
+use bisync::synchronous::*;
+
+#[cfg(feature = "serial2")]
+/// Public re-export of the serial2 crate.
+pub use serial2;
+#[cfg(feature = "serial2")]
+use serial2::SerialPort as Serial2Port;
+
 pub mod instructions;
 
 mod client;
