+++ src/serial_port/mod.rs
@@ -4,10 +4,13 @@ use core::time::Duration;
 
 #[cfg(feature = "serial2")]
 pub mod serial2;
+#[cfg(feature = "serial2")]
+use super::Serial2Port;
 
 /// [`SerialPort`]s are used to communicate with the hardware by reading and writing data.
 ///
 /// The implementor of the trait must also configure the serial line to use 8 bits characters, 1 stop bit, no parity and no flow control.
+#[super::bisync]
 pub trait SerialPort {
 	/// The error type returned by the serial port when reading, writing or setting the baud rate.
 	type Error;
@@ -25,10 +28,12 @@ pub trait SerialPort {
 	fn discard_input_buffer(&mut self) -> Result<(), Self::Error>;
 
 	/// Returns available bytes to read, blocking until at least one byte is available or the deadline expires.
-	fn read(&mut self, buffer: &mut [u8], deadline: &Self::Instant) -> Result<usize, Self::Error>;
+	#[allow(async_fn_in_trait)]
+	async fn read(&mut self, buffer: &mut [u8], deadline: &Self::Instant) -> Result<usize, Self::Error>;
 
 	/// Write all bytes in the buffer to the serial port.
-	fn write_all(&mut self, buffer: &[u8]) -> Result<(), Self::Error>;
+	#[allow(async_fn_in_trait)]
+	async fn write_all(&mut self, buffer: &[u8]) -> Result<(), Self::Error>;
 
 	/// Make a deadline to expire after the given timeout.
 	fn make_deadline(&self, timeout: Duration) -> Self::Instant;
